+++ speeduino/decoders.h
@@ -16,7 +16,7 @@ void loggerPrimaryISR();
 void loggerSecondaryISR();
 static inline uint16_t stdGetRPM(uint16_t);
 static inline void setFilter(unsigned long);
-static inline int crankingGetRPM(uint8_t);
+static inline int16_t crankingGetRPM(uint8_t);
 //static inline void doPerToothTiming(uint16_t);
 
 //All of the below are the 6 required functions for each decoder / pattern
@@ -24,134 +24,134 @@ void triggerSetup_missingTooth();
 void triggerPri_missingTooth();
 void triggerSec_missingTooth();
 uint16_t getRPM_missingTooth();
-int getCrankAngle_missingTooth();
+int16_t getCrankAngle_missingTooth();
 extern void triggerSetEndTeeth_missingTooth();
 
 void triggerSetup_DualWheel();
 void triggerPri_DualWheel();
 void triggerSec_DualWheel();
 uint16_t getRPM_DualWheel();
-int getCrankAngle_DualWheel();
+int16_t getCrankAngle_DualWheel();
 void triggerSetEndTeeth_DualWheel();
 
 void triggerSetup_BasicDistributor();
 void triggerPri_BasicDistributor();
 void triggerSec_BasicDistributor();
 uint16_t getRPM_BasicDistributor();
-int getCrankAngle_BasicDistributor();
+int16_t getCrankAngle_BasicDistributor();
 void triggerSetEndTeeth_BasicDistributor();
 
 void triggerSetup_GM7X();
 void triggerPri_GM7X();
 void triggerSec_GM7X();
 uint16_t getRPM_GM7X();
-int getCrankAngle_GM7X();
+int16_t getCrankAngle_GM7X();
 void triggerSetEndTeeth_GM7X();
 
 void triggerSetup_4G63();
 void triggerPri_4G63();
 void triggerSec_4G63();
 uint16_t getRPM_4G63();
-int getCrankAngle_4G63();
+int16_t getCrankAngle_4G63();
 void triggerSetEndTeeth_4G63();
 
 void triggerSetup_24X();
 void triggerPri_24X();
 void triggerSec_24X();
 uint16_t getRPM_24X();
-int getCrankAngle_24X();
+int16_t getCrankAngle_24X();
 void triggerSetEndTeeth_24X();
 
 void triggerSetup_Jeep2000();
 void triggerPri_Jeep2000();
 void triggerSec_Jeep2000();
 uint16_t getRPM_Jeep2000();
-int getCrankAngle_Jeep2000();
+int16_t getCrankAngle_Jeep2000();
 void triggerSetEndTeeth_Jeep2000();
 
 void triggerSetup_Audi135();
 void triggerPri_Audi135();
 void triggerSec_Audi135();
 uint16_t getRPM_Audi135();
-int getCrankAngle_Audi135();
+int16_t getCrankAngle_Audi135();
 void triggerSetEndTeeth_Audi135();
 
 void triggerSetup_HondaD17();
 void triggerPri_HondaD17();
 void triggerSec_HondaD17();
 uint16_t getRPM_HondaD17();
-int getCrankAngle_HondaD17();
+int16_t getCrankAngle_HondaD17();
 void triggerSetEndTeeth_HondaD17();
 
 void triggerSetup_Miata9905();
 void triggerPri_Miata9905();
 void triggerSec_Miata9905();
 uint16_t getRPM_Miata9905();
-int getCrankAngle_Miata9905();
+int16_t getCrankAngle_Miata9905();
 void triggerSetEndTeeth_Miata9905();
-int getCamAngle_Miata9905();
+int16_t getCamAngle_Miata9905();
 
 void triggerSetup_MazdaAU();
 void triggerPri_MazdaAU();
 void triggerSec_MazdaAU();
 uint16_t getRPM_MazdaAU();
-int getCrankAngle_MazdaAU();
+int16_t getCrankAngle_MazdaAU();
 void triggerSetEndTeeth_MazdaAU();
 
 void triggerSetup_non360();
 void triggerPri_non360();
 void triggerSec_non360();
 uint16_t getRPM_non360();
-int getCrankAngle_non360();
+int16_t getCrankAngle_non360();
 void triggerSetEndTeeth_non360();
 
 void triggerSetup_Nissan360();
 void triggerPri_Nissan360();
 void triggerSec_Nissan360();
 uint16_t getRPM_Nissan360();
-int getCrankAngle_Nissan360();
+int16_t getCrankAngle_Nissan360();
 void triggerSetEndTeeth_Nissan360();
 
 void triggerSetup_Subaru67();
 void triggerPri_Subaru67();
 void triggerSec_Subaru67();
 uint16_t getRPM_Subaru67();
-int getCrankAngle_Subaru67();
+int16_t getCrankAngle_Subaru67();
 void triggerSetEndTeeth_Subaru67();
 
 void triggerSetup_Daihatsu();
 void triggerPri_Daihatsu();
 void triggerSec_Daihatsu();
 uint16_t getRPM_Daihatsu();
-int getCrankAngle_Daihatsu();
+int16_t getCrankAngle_Daihatsu();
 void triggerSetEndTeeth_Daihatsu();
 
 void triggerSetup_Harley();
 void triggerPri_Harley();
 void triggerSec_Harley();
 uint16_t getRPM_Harley();
-int getCrankAngle_Harley();
+int16_t getCrankAngle_Harley();
 void triggerSetEndTeeth_Harley();
 
 void triggerSetup_ThirtySixMinus222();
 void triggerPri_ThirtySixMinus222();
 void triggerSec_ThirtySixMinus222();
 uint16_t getRPM_ThirtySixMinus222();
-int getCrankAngle_ThirtySixMinus222();
+int16_t getCrankAngle_ThirtySixMinus222();
 void triggerSetEndTeeth_ThirtySixMinus222();
 
 void triggerSetup_ThirtySixMinus21();
 void triggerPri_ThirtySixMinus21();
 void triggerSec_ThirtySixMinus21();
 uint16_t getRPM_ThirtySixMinus21();
-int getCrankAngle_ThirtySixMinus21();
+int16_t getCrankAngle_ThirtySixMinus21();
 void triggerSetEndTeeth_ThirtySixMinus21();
 
 void triggerSetup_420a();
 void triggerPri_420a();
 void triggerSec_420a();
 uint16_t getRPM_420a();
-int getCrankAngle_420a();
+int16_t getCrankAngle_420a();
 void triggerSetEndTeeth_420a();
 
 void triggerPri_Webber();
@@ -160,7 +160,7 @@ void triggerSec_Webber();
 extern void (*triggerHandler)(); //Pointer for the trigger function (Gets pointed to the relevant decoder)
 extern void (*triggerSecondaryHandler)(); //Pointer for the secondary trigger function (Gets pointed to the relevant decoder)
 extern uint16_t (*getRPM)(); //Pointer to the getRPM function (Gets pointed to the relevant decoder)
-extern int (*getCrankAngle)(); //Pointer to the getCrank Angle function (Gets pointed to the relevant decoder)
+extern int16_t (*getCrankAngle)(); //Pointer to the getCrank Angle function (Gets pointed to the relevant decoder)
 extern void (*triggerSetEndTeeth)(); //Pointer to the triggerSetEndTeeth function of each decoder
 
 extern volatile unsigned long curTime;
@@ -187,7 +187,7 @@ extern volatile unsigned long toothOneTime; //The time (micros()) that tooth 1 l
 extern volatile unsigned long toothOneMinusOneTime; //The 2nd to last time (micros()) that tooth 1 last triggered
 extern volatile bool revolutionOne; // For sequential operation, this tracks whether the current revolution is 1 or 2 (not 1)
 
-extern volatile unsigned int secondaryToothCount; //Used for identifying the current secondary (Usually cam) tooth for patterns with multiple secondary teeth
+extern volatile uint16_t secondaryToothCount; //Used for identifying the current secondary (Usually cam) tooth for patterns with multiple secondary teeth
 extern volatile unsigned long secondaryLastToothTime; //The time (micros()) that the last tooth was registered (Cam input)
 extern volatile unsigned long secondaryLastToothTime1; //The time (micros()) that the last tooth was registered (Cam input)
 
@@ -195,7 +195,7 @@ extern volatile uint16_t triggerActualTeeth;
 extern volatile unsigned long triggerFilterTime; // The shortest time (in uS) that pulses will be accepted (Used for debounce filtering)
 extern volatile unsigned long triggerSecFilterTime; // The shortest time (in uS) that pulses will be accepted (Used for debounce filtering) for the secondary input
 extern volatile bool validTrigger; //Is set true when the last trigger (Primary or secondary) was valid (ie passed filters)
-extern unsigned int triggerSecFilterTime_duration; // The shortest valid time (in uS) pulse DURATION
+extern uint16_t triggerSecFilterTime_duration; // The shortest valid time (in uS) pulse DURATION
 extern volatile uint16_t triggerToothAngle; //The number of crank degrees that elapse per tooth
 extern volatile bool triggerToothAngleIsCorrect; //Whether or not the triggerToothAngle variable is currently accurate. Some patterns have times when the triggerToothAngle variable cannot be accurately set.
 extern bool secondDerivEnabled; //The use of the 2nd derivative calculation is limited to certain decoders. This is set to either true or false in each decoders setup routine
